Develop merge

+++ server/src/questions/questions.service.ts
@@ -1,11 +1,11 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { HttpException, Inject, Injectable } from '@nestjs/common';
 import { Model } from 'mongoose';
 import { Question, Questions } from 'src/interfaces/questions.interface';
 import {
   AppendQuestionsDto,
   CreateQuestionDto,
 } from '../dto/create-question.dto';
-
+import { DifficultyVoteDto } from './dto/difficulty-vote.dto';
 @Injectable()
 export class QuestionsService {
   constructor(
@@ -46,6 +46,7 @@ export class QuestionsService {
     if (questionSet.author.toString() !== user._id.toString()) {
       throw new Error('You are not the author of this question set');
     }
+    console.log(appendQuestionsDto.questions);
     // Ensure questions is an array
     const questions = await this.questionModel.create(
       appendQuestionsDto.questions,
@@ -55,16 +56,54 @@ export class QuestionsService {
     await questionSet.save();
     return questionsArray;
   }
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   async deleteOne(id: string, userId: string): Promise<Question> {
-    //TODO check if user is author
+    // TODO this is inefficient
+    const question = await this.questionModel.findById(id);
+    if (!question) {
+      throw new HttpException('Question not found', 404);
+    }
+    const questionSet = await this.questionsSetsModel.findOne({
+      questions: id,
+    });
+    if (!questionSet) {
+      throw new HttpException('Question not found in any set', 404);
+    }
+    if (questionSet.author.toString() !== userId) {
+      throw new HttpException(
+        'You are not the author of this question set',
+        403,
+      );
+    }
     await this.questionsSetsModel
       .updateMany({ questions: id }, { $pull: { questions: id } })
       .exec();
     return this.questionModel.findByIdAndDelete(id).exec();
   }
-  async updateOne(id: string, body: CreateQuestionDto): Promise<Question> {
-    //TODO check if user is author
+  async getOne(id: string): Promise<Question> {
+    return this.questionModel.findById(id).exec();
+  }
+  async updateOne(
+    id: string,
+    body: CreateQuestionDto,
+    userId: string,
+  ): Promise<Question> {
+    //TODO
+    const question = await this.questionModel.findById(id);
+    if (!question) {
+      throw new HttpException('Question not found', 404);
+    }
+    const questionSet = await this.questionsSetsModel.findOne({
+      questions: id,
+    });
+    if (!questionSet) {
+      throw new HttpException('Question not found in any set', 404);
+    }
+    if (questionSet.author.toString() !== userId) {
+      throw new HttpException(
+        'You are not the author of this question set',
+        403,
+      );
+    }
     const updatedQuestion = await this.questionModel
       .findByIdAndUpdate(id, body, { new: true })
       .exec();
@@ -95,5 +134,41 @@ export class QuestionsService {
       deleted: deleted.deletedCount,
       found: questionsNotInAnySet.length,
     };
+  }
+  async setExplanation(id: string, explanation: string): Promise<string> {
+    const question = await this.questionModel.findById(id);
+    question.explanation = explanation;
+    await question.save();
+    return explanation;
+  }
+  async incrementReport(id: string): Promise<number> {
+    const question = await this.questionModel.findById(id);
+    question.report++;
+    await question.save();
+    return question.report;
+  }
+  async voteDifficulty(
+    questionId: string,
+    DifficultyVoteDto: DifficultyVoteDto,
+    userId: string,
+  ) {
+    const question = await this.questionModel.findById(questionId);
+    const user = await this.usersModel.findById(userId);
+    if (!user) {
+      throw new HttpException('User not found', 404);
+    }
+    if (!question) {
+      throw new HttpException('Question not found', 404);
+    }
+    // Check if user has already voted
+    if (question.difficulty.some((vote) => vote.user === userId)) {
+      // Remove previous vote
+      question.difficulty = question.difficulty.filter(
+        (vote) => vote.user !== userId,
+      );
+    }
+    question.difficulty.push({ user: userId, value: DifficultyVoteDto.value });
+    await question.save();
+    return question.difficulty;
   }
 }
